# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentAssetBody": {
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "role": {
            "pattern": "^(top|side|supporting)$",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "role"
        ],
        "title": "AssessmentAssetBody",
        "type": "object"
      },
...
      "BodyAssessmentBody": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          },
          "assessed_at": {
            "format": "date-time",
            "title": "Assessed At",
            "type": "string"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssessmentAssetBody"
            },
            "maxItems": 10,
            "title": "Assets",
            "type": "array"
          },
          "bcs_scale": {
            "default": 9,
            "maximum": 9.0,
            "minimum": 9.0,
            "title": "Bcs Scale",
            "type": "integer"
          },
          "bcs_score": {
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Bcs Score",
            "type": "integer"
          },
          "muscle_condition": {
            "pattern": "^(normal|mild_loss|moderate_loss|severe_loss|unknown)$",
            "title": "Muscle Condition",
            "type": "string"
          }
        },
        "required": [
          "bcs_score",
          "muscle_condition",
          "answers",
          "assessed_at"
        ],
        "title": "BodyAssessmentBody",
        "type": "object"
      },
...
      "ConsentBody": {
        "properties": {
          "policy_version": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Policy Version",
            "type": "string"
          },
          "purpose": {
            "pattern": "^(body_photographs|medical_records)$",
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "policy_version"
        ],
        "title": "ConsentBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "IdResponse",
        "type": "object"
      },
...
      "PetBody": {
        "properties": {
          "birth_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "species": {
            "pattern": "^(dog|cat)$",
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "name",
          "species"
        ],
        "title": "PetBody",
        "type": "object"
      },
...
      "PetSummary": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "species": {
            "enum": [
              "dog",
              "cat"
            ],
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "species"
        ],
        "title": "PetSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Pet Platform Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/pets/{pet_id}": {
      "get": {
        "operationId": "pet_knowledge_api_v1_knowledge_pets__pet_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pet Knowledge Api V1 Knowledge Pets  Pet Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pet Knowledge",
        "tags": [
          "pet-knowledge"
        ]
      }
    },
...
    "/api/v1/pet-life/households/{household_id}/pets": {
      "get": {
        "operationId": "list_household_pets_api_v1_pet_life_households__household_id__pets_get",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PetSummary"
                  },
                  "title": "Response List Household Pets Api V1 Pet Life Households  Household Id  Pets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Household Pets",
        "tags": [
          "pet-life"
        ]
      },
      "post": {
        "operationId": "create_pet_api_v1_pet_life_households__household_id__pets_post",
        "parameters": [
          {
            "in": "path",
            "name": "household_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Household Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Pet",
        "tags": [
          "pet-life"
        ]
      }
    },
...
    "/api/v1/pet-life/pets/{pet_id}/assets": {
      "get": {
        "operationId": "list_pet_assets_api_v1_pet_life_pets__pet_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Pet Assets Api V1 Pet Life Pets  Pet Id  Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pet Assets",
        "tags": [
          "pet-health-assets"
        ]
      },
      "post": {
        "operationId": "upload_pet_asset_api_v1_pet_life_pets__pet_id__assets_post",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Filename",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "X-Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Asset-Category",
            "required": true,
            "schema": {
              "pattern": "^(body_top|body_side|medical_document|lab_result|other_medical)$",
              "title": "X-Asset-Category",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Consent-ID",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Consent-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Captured-At",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Captured-At"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Pet Asset Api V1 Pet Life Pets  Pet Id  Assets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Pet Asset",
        "tags": [
          "pet-health-assets"
        ]
      }
    },
    "/api/v1/pet-life/pets/{pet_id}/assets/{asset_id}": {
      "delete": {
        "operationId": "remove_pet_asset_api_v1_pet_life_pets__pet_id__assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Pet Asset",
        "tags": [
          "pet-health-assets"
        ]
      },
      "get": {
        "operationId": "download_pet_asset_api_v1_pet_life_pets__pet_id__assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Pet Asset",
        "tags": [
          "pet-health-assets"
        ]
      }
    },
    "/api/v1/pet-life/pets/{pet_id}/body-assessments": {
      "get": {
        "operationId": "list_body_assessments_api_v1_pet_life_pets__pet_id__body_assessments_get",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Body Assessments Api V1 Pet Life Pets  Pet Id  Body Assessments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Body Assessments",
        "tags": [
          "pet-health-assets"
        ]
      },
      "post": {
        "operationId": "create_body_assessment_api_v1_pet_life_pets__pet_id__body_assessments_post",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BodyAssessmentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Body Assessment Api V1 Pet Life Pets  Pet Id  Body Assessments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Body Assessment",
        "tags": [
          "pet-health-assets"
        ]
      }
    },
...
    "/api/v1/pet-life/pets/{pet_id}/consents": {
      "post": {
        "operationId": "grant_pet_consent_api_v1_pet_life_pets__pet_id__consents_post",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grant Pet Consent Api V1 Pet Life Pets  Pet Id  Consents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grant Pet Consent",
        "tags": [
          "pet-health-assets"
        ]
      }
    },
    "/api/v1/pet-life/pets/{pet_id}/consents/{consent_id}/withdraw": {
      "post": {
        "operationId": "withdraw_pet_consent_api_v1_pet_life_pets__pet_id__consents__consent_id__withdraw_post",
        "parameters": [
          {
            "in": "path",
            "name": "pet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pet Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "consent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Consent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw Pet Consent",
        "tags": [
          "pet-health-assets"
        ]
      }
    },
...
  }
}